Delete log dir only if it exists

+++ spec/spec_helper.rb
@@ -23,6 +23,6 @@ RSpec.configure do |config|
     # TODO: Find out what causes logger to create this dir when running specs.
     #       There's probably a test app class being created somewhere with root
     #       not pointing to a tmp dir.
-    FileUtils.rm_r(LOG_DIR)
+    FileUtils.rm_r(LOG_DIR) if LOG_DIR.exist?
   end
 end
